remove unused assets, utils, tsx files
